Custom List Scores Dispalying & Editing
- Custom score types are now supported based on user preferences
- Many other changes related to scoring types

+++ app/src/main/java/com/mxt/anitrend/model/entity/anilist/MediaList.java
@@ -192,7 +192,7 @@ public class MediaList extends RecyclerItem implements Parcelable, Cloneable {
         this.status = status;
     }
 
-    public void setScore(int score) {
+    public void setScore(float score) {
         this.score = score;
     }
